backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
alembic==1.15.2
pydantic-settings==2.9.1
google-api-python-client==2.166.0
google-auth-oauthlib==1.2.1
google-auth==2.38.0
beautifulsoup4==4.13.3
